Completed CPU implementation.

+++ src/data_structures/instruction_cache.c
@@ -9,11 +9,11 @@ int instructionCacheController[64];
 // Bit 4-9: address of row
 // Bit 10-13: address of instruction in row.
 // So 64 rows of 64 bytes (each row of 64 bytes is broken into 8 sections of 8 bytes each
-double instructionCacheData[64][8];
+int64_t instructionCacheData[64][8];
 
-double fetchInstruction(int programCounter) {
+int64_t fetchInstruction(int programCounter) {
     int controllerData = instructionCacheController[programCounter];
-    if((controllerData >> 15) == 0) {
+    if((controllerData >> 15) == 0) { // Check if record is valid.
         return ERR; // Cache miss
     }
     int rowMask = 0b0001111110000000;
@@ -23,8 +23,8 @@ double fetchInstruction(int programCounter) {
     return instructionCacheData[row][cell];
 }
 
-void loadInstructionCache(int programCounter, double instruction) {
-    int cacheMask = 0b100 << 13; // Mask this to the top order bits.
+void loadInstructionCache(int programCounter, int64_t instruction) {
+    int cacheMask = 0b100 << 13; // Mask valid/dirty/lru to the top order bits.
     int rowMask = programCounter / 8 << 7; // Mask this to bits 4-9 for column
     int instructionMask = programCounter % 8 << 3; // Mask bits for cell
     int instructionCacheAddress = cacheMask | rowMask | instructionMask;
